MDL-17524 Merging from HEAD Deleting a calculated question delete datasetdefinitions shared by other question

+++ question/type/calculated/questiontype.php
@@ -468,8 +468,13 @@ class question_calculated_qtype extends question_dataset_dependent_questiontype 
         delete_records("question_numerical_units", "question", $questionid);
         if ($datasets = get_records('question_datasets', 'question', $questionid)) {
             foreach ($datasets as $dataset) {
-                delete_records('question_dataset_definitions', 'id', $dataset->datasetdefinition);
-                delete_records('question_dataset_items', 'definition', $dataset->datasetdefinition);
+                if (! get_records_select(
+                        'question_datasets',
+                        "question != $questionid
+                        AND datasetdefinition = $dataset->datasetdefinition;")){                                 
+                    delete_records('question_dataset_definitions', 'id', $dataset->datasetdefinition);
+                    delete_records('question_dataset_items', 'definition', $dataset->datasetdefinition);
+                }
             }
         }
         delete_records("question_datasets", "question", $questionid);
